requirements.txt
torch>=1.7.1
torchvision>=0.8.2
torchaudio>=0.7.2
transformers>=4.5.0
Pillow>=8.0.0
opencv-python>=4.5.0
tqdm>=4.60.0
numpy>=1.19.0
pandas>=1.2.0
rdflib>=6.0.0
requests>=2.25.0
gdown>=4.7.1
SPARQLWrapper>=2.0.0 